src/mltest/utils.py
import os
import sys
from src.mltest.exception import CustomException
from src.mltest.logger import logging
import pandas as pd
from dotenv import load_dotenv
import pymysql

load_dotenv()

host = os.getenv("host")
user = os.getenv("user")    
password = os.getenv("password")
db = os.getenv("db")


def read_sql_data():
    logging.info("Reading data from SQL database")
    try:
        mydb=pymysql.connect(
            host=host,
            user=user,
            password=password,
            db=db 
        )
        logging.info(f"Database connection established successfully: {mydb}")
        df=pd.read_sql_query("select * from bmw",mydb)
        print(df.head())

        return df

    except Exception as ex:
        raise CustomException(ex)