requirements.txt
numpy
pandas
mysql-connector-python
python-dotenv
pymysql
scikit-learn